refactor(carousel): consolidate and relocate type definitions
- Removed `types.ts` and `types.test.tsx` from `carousel/variants` and merged their contents into `carousel`.
- Updated imports in all related components (`CarouselArrowIcons.tsx`, `CarouselArrowIndex.tsx`, `CarouselNavigationArrows.tsx`, `CarouselDots.tsx`, `CarouselArrows.tsx`) to use the new consolidated `types.ts`.
- Restored and adjusted tests for `CarouselWrapperProps` under the new structure.
