+++ src/components/carousel/types.ts
@@ -4,6 +4,9 @@ import {IconifyProps} from "@src/components/iconify";
 import {StackProps} from "@mui/material/Stack";
 import React from "react";
 import {IconButtonProps} from "@mui/material/IconButton";
+import {Profile} from "@lens-protocol/api-bindings";
+// @ts-ignore
+import type {Post} from "@lens-protocol/api-bindings/dist/declarations/src/lens/graphql/generated";
 
 export interface CarouselDotsProps extends BoxProps {
   rounded?: boolean;
@@ -50,3 +53,42 @@ export type CarouselDotsStyledRootProps = {
   rounded: boolean;
 };
 
+export interface CarouselWrapperProps<T> {
+  data: T[];
+  title?: string;
+  minItemWidth: number;
+  maxItemWidth: number;
+  renderSlide: (slideItems: T[], itemsPerRow: number, index: number) => React.ReactNode;
+  carouselSettings: any;
+  boxStyle?: any;
+  boxClassName?: string;
+}
+
+export type CarouselSlideProps = {
+  items: Profile[];
+  itemsPerRow: number;
+};
+
+export type CarouselCreatorsProps = {
+  data: Profile[];
+  title?: string;
+  minItemWidth: number;
+  maxItemWidth: number;
+};
+
+export type CarouselPosterMiniProps = {
+  data: Post[];
+  title?: string;
+  minItemWidth: number;
+  maxItemWidth: number;
+};
+
+export type CarouselTopTitlesProps = {
+  posts: Post[];
+  category?: string;
+};
+
+export type CarouselPosterSlideProps = {
+  items: Post[];
+  itemsPerRow: number;
+};
